+++ src/components/EditAlarm.vue
@@ -1,29 +1,15 @@
 <template>
 	<v-col cols='12' class=''>
-		<v-row justify='center' class='ma-0 pa-0 mb-n6' >
+		<v-row justify='center' class='ma-0 pa-0 mb-n6'>
 			<v-col cols='12' class='ma-0 pa-0'>
 				<v-row justify='center' class='ma-0 pa-0'>
 					<v-col cols='5'>
-						<v-select
-							v-model='chosenHour'
-							:items='hours'
-							bg-color='offwhite'
-							color='primary'
-							density='compact'
-							label='Select Hour'
-							variant='outlined'
-						/>
+						<v-select v-model='chosenHour' :items='hours' bg-color='offwhite' color='primary'
+							density='compact' label='Select Hour' variant='outlined' />
 					</v-col>
 					<v-col cols='5'>
-						<v-select
-							v-model='chosenMinute'
-							:items='minutes'
-							bg-color='offwhite'
-							color='primary'
-							density='compact'
-							label='Select Minute'
-							variant='outlined'
-						/>
+						<v-select v-model='chosenMinute' :items='minutes' bg-color='offwhite' color='primary'
+							density='compact' label='Select Minute' variant='outlined' />
 					</v-col>
 				</v-row>
 			</v-col>
@@ -31,30 +17,19 @@
 		<v-row justify='center' align='center' class='ma-0 pa-0'>
 
 			<v-col cols='auto'>
-				<v-btn
-					@click='alarm_delete'
-					:disabled='!currentAlarm'
-					color='error'
-					rounded='lg'
-					size='small'
-					variant='elevated'
-				>
+				<v-btn @click='alarm_delete' :disabled='!currentAlarm' color='error' rounded='lg' size='small'
+					variant='elevated'>
 					<span class='text-white'>delete alarm</span>
 					<v-icon style='vertical-align: middle;' class='ml-2' size='small' :icon='mdiClose' color='white' />
 				</v-btn>
 			</v-col>
 
 			<v-col cols='auto'>
-				<v-btn
-					@click='addAlarm'
-					:disabled='disabled'
-					color='secondary'
-					rounded='lg'
-					size='small'
-					variant='elevated'
-				>
-					<v-icon style='vertical-align: middle;' class='mr-2' size='small' :icon='mdiAlarmPlus' color='white' />
-					<span >{{ buttonAlarmText }} alarm</span>
+				<v-btn @click='addAlarm' :disabled='disabled' color='secondary' rounded='lg' size='small'
+					variant='elevated'>
+					<v-icon style='vertical-align: middle;' class='mr-2' size='small' :icon='mdiAlarmPlus'
+						color='white' />
+					<span>{{ buttonAlarmText }} alarm</span>
 				</v-btn>
 			</v-col>
 		</v-row>
@@ -67,37 +42,33 @@ import { zeroPad } from '@/vanillaTS/zeropad';
 import type { nu } from '@/types';
 import type { Ref } from 'vue';
 
-const [ statusStore, websocketStore ] = [ statusModule(), websocketModule() ];
+const [statusStore, websocketStore] = [statusModule(), websocketModule()];
 
-const emit = defineEmits([ 'close' ]);
+const emit = defineEmits(['close']);
 
-const buttonAlarmText = computed((): string => {
-	return statusStore.alarm ? 'update' : 'add';
-});
+const buttonAlarmText = computed(() => statusStore.alarm ? 'update' : 'add');
 
-const currentAlarm = computed(() => {
-	return statusStore.alarm;
-});
+const currentAlarm = computed(() => statusStore.alarm);
 
 const hours = computed((): Array<{
 	title: string;
-	value: number; 
+	value: number;
 }> => {
 	const hours = [];
-	for (const [ index, _item ] of new Array(24).entries()) hours.push({
+	for (const [index, _item] of new Array(24).entries()) hours.push({
 		title: zeroPad(index),
-		value: index 
+		value: index
 	});
 	return hours;
 });
 const minutes = computed((): Array<{
 	title: string;
-	value: number; 
+	value: number;
 }> => {
 	const minutes = [];
-	for (const [ index, _item ] of new Array(60).entries()) minutes.push({
+	for (const [index, _item] of new Array(60).entries()) minutes.push({
 		title: zeroPad(index),
-		value: index 
+		value: index
 	});
 	return minutes;
 });
@@ -105,26 +76,24 @@ const minutes = computed((): Array<{
 const chosenHour: Ref<nu> = ref(undefined);
 const chosenMinute: Ref<nu> = ref(undefined);
 
-const disabled = computed((): boolean => {
-	return !(chosenHour.value !== undefined && chosenHour.value >= 0 && chosenMinute.value !== undefined && chosenMinute.value >= 0);
-});
+const disabled = computed(() => !(chosenHour.value !== undefined && chosenHour.value >= 0 && chosenMinute.value !== undefined && chosenMinute.value >= 0));
 const addAlarm = (): void => {
 	if (chosenHour.value === undefined) return;
 	if (chosenMinute.value === undefined) return;
 	loadingModule().set_loading(true);
 	const body = {
 		hour: chosenHour.value,
-		minute: chosenMinute.value 
+		minute: chosenMinute.value
 	};
 	if (statusStore.alarm) {
 		websocketStore.send({
 			name: 'alarm_update',
-			body 
+			body
 		});
 	} else {
 		websocketStore.send({
 			name: 'alarm_add',
-			body 
+			body
 		});
 	}
 	emit('close');