+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# <a href='https://github.com/mrjackwills/push_alarm_frontend/releases/tag/v0.3.1'>v0.3.1</a>
+### 2025-04-07
+
+### Chores
++ dependencies updated, [06941f5d](https://github.com/mrjackwills/push_alarm_frontend/commit/06941f5de683acd440e996ade18752454714dea4)
+
+### Refactors
++ computed() methods, [79a00b38](https://github.com/mrjackwills/push_alarm_frontend/commit/79a00b3834edae33bb21c59043dac3682543e1fb)
+
 # <a href='https://github.com/mrjackwills/push_alarm_frontend/releases/tag/v0.3.0'>v0.3.0</a>
 ### 2025-02-25